requirements.txt
langchain==0.1.0
langchain-anthropic==0.0.9
playwright==1.41.2
python-dotenv==1.0.0
pyyaml==6.0.1
validators==0.22.0
tenacity==8.2.3
pillow==10.2.0 